Added manual projectile clean up
Added manual clean up of projectiles.
Tag added to projectiles so that they can be found and cleaned.
DestroyFireball changed to DestroyProjectile.
Just in case we ever consolidate projectiles to their own class.
Moved arena back because a lot of the cover was disappearing on load.
When you reload a scene unity ignores the order and places them based on
layer or z location >=[