Bunch of stuff. Read description please
1. Input controller will now pause the game and prompt the user to
reconnect controller.
2. The way the thing works for now, is that the game will assume you
have as many controllers connected as you did when you first started
this script. It will assign the other players to keyboard controls.
3. The keyboard controls for all the other players is temporary until we
make automatic scene population and respawn points.
4. I needed to get this really quick so that we can just plug and play
and not really have to assign controllers manually.
5. Gamemode code is a work in progress, I was trying to fix the bug that
makes it so that if you press the right trigger of a ps4 controller, and
there is an xbox controller plugged it, then the pause button of the
xbox controller gets recognized.
5. This is a pain in the ass and its a known problem with unity. I was
trying to find some funky way around it by using a list of ps4
controllers and what not. BUT FUCK IT.
6. This list is not really a list. I don't give a fuck.
7. Also, I just realize that my branch name makes no sense. My
apologies. It should be called, assMagic or something.